+++ codbex-invoices/events/generate/CreditNote/index.html
@@ -6,11 +6,10 @@
         <meta name="viewport" content="width=device-width, initial-scale=1">
         <link rel="icon" sizes="any" href="data:;base64,iVBORw0KGgo=">
         <title dg-view-title></title>
-        <script type="text/javascript" src="credit-note-generate.js"></script>
+        <script type="text/javascript" src="./config.js"></script>
         <script src="/services/js/platform-core/services/loader.js?id=view-js"></script>
         <link rel="stylesheet" href="/services/js/platform-core/services/loader.js?id=view-css" />
-        <theme></theme>
-        <script type="text/javascript" src="generate-credit-note-controller.js">
+        <script type="text/javascript" src="./controller.js">
         </script>
     </head>
