+++ codbex-invoices/events/generate/DebitNote/config.js
@@ -1,12 +1,11 @@
 const viewData = {
     id: 'debit-note-generate',
     label: 'Generate Debit Note',
-    path: '/services/web/codbex-invoices/events/generate/DebitNote/generate-debit-note.html',
-    lazyLoad: false,
-    autoFocusTab: false,
-    maxWidth: '320px',
-    maxHeight: '182px',
-    perspective: 'salesinvoice',
+    path: '/services/web/codbex-invoices/events/generate/DebitNote/index.html',
+    translation: {
+        key: 'codbex-invoices:t.DEBITNOTE',
+    },
+    perspective: 'SalesInvoice',
     view: 'SalesInvoice',
     type: 'entity',
     order: 9
